리팩토링: Carousel 컴포넌트 로직을 useCarousel 훅으로 분리
