+++ hooks/ui/index.ts
@@ -7,3 +7,5 @@ export { useToast } from './useToast';
 export { useDropdown } from './useDropdown';
 
 export { default as useSelect } from './useSelect';
+
+export { useCarousel } from './useCarousel';
